# src/album/album.controller.ts
import {
  Controller,
  Get,
  Post,
  Put,
  Delete,
  Body,
  Param,
  HttpCode,
  ParseUUIDPipe,
  HttpStatus,
} from '@nestjs/common';
import { AlbumService } from './album.service';
import { CreateAlbumDto } from './dto/create-album.dto';
import { UpdateAlbumDto } from './dto/update-album.dto';
import { LoggingService } from 'src/logger/logger.service';

@Controller('album')
export class AlbumController {
  constructor(
    private readonly albumService: AlbumService,
    private readonly logger: LoggingService,
  ) {}

  @Get()
  findAll() {
    this.logger.verbose('[AlbumController] findAll called');
    return this.albumService.findAll();
  }

  @Get(':id')
  findOne(@Param('id', ParseUUIDPipe) id: string) {
    this.logger.verbose(`[AlbumController] findOne called with id: ${id}`);
    return this.albumService.findOne(id);
  }

  @Post()
  @HttpCode(HttpStatus.CREATED)
  create(@Body() createAlbumDto: CreateAlbumDto) {
    this.logger.verbose(
      `[AlbumController] create called with data: ${JSON.stringify(createAlbumDto)}`,
    );
    return this.albumService.create(createAlbumDto);
  }

  @Put(':id')
  update(
    @Param('id', ParseUUIDPipe) id: string,
    @Body() updateAlbumDto: UpdateAlbumDto,
  ) {
    this.logger.verbose(
      `[AlbumController] update called with id: ${id} and data: ${JSON.stringify(updateAlbumDto)}`,
    );
    return this.albumService.update(id, updateAlbumDto);
  }

  @Delete(':id')
  @HttpCode(HttpStatus.NO_CONTENT)
  remove(@Param('id', ParseUUIDPipe) id: string) {
    this.logger.verbose(`[AlbumController] remove called with id: ${id}`);
    return this.albumService.remove(id);
  }
}